// packages/core/src/generated-core/models/Song.ts
/**
 * Operation real wave API
 * MSR Soundtrack audio management website\'s schema
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { FileStatus } from '../models/FileStatus.ts';
import { HttpFile } from '../http/http.ts';

export class Song {
    'cid': string;
    'name': string;
    'artistes': Array<string>;
    'status': FileStatus;
    'filePath': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "cid",
            "baseName": "cid",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "artistes",
            "baseName": "artistes",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "FileStatus",
            "format": ""
        },
        {
            "name": "filePath",
            "baseName": "filePath",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Song.attributeTypeMap;
    }

    public constructor() {
    }
}